java: globally use double instead of float

+++ java/src/grains_lxc/svm/Demo.java
@@ -4,7 +4,6 @@
  */
 package grains_lxc.svm;
 
-import java.util.ArrayList;
 import java.util.List;
 import grains_lxc.features.My_Features;
 
@@ -19,23 +18,23 @@ public class Demo {
 		features_train.load_saved_features();
 		features_test.load_saved_features();
 		
-		List<List<Float>> train_x=features_train.get_features_x();
-		List<Float> train_y=features_train.get_features_y();
+		List<List<Double>> train_x=features_train.get_features_x();
+		List<Double> train_y=features_train.get_features_y();
 		int train_number=train_y.size();
 		
-		List<List<Float>> test_x=features_test.get_features_x();
-		List<Float> test_y=features_test.get_features_y();
+		List<List<Double>> test_x=features_test.get_features_x();
+		List<Double> test_y=features_test.get_features_y();
 		int test_number=test_x.size();
 
 		// 训练，预测
 		my_svm.train(train_x, train_y);
-		List<Float> predict_y=my_svm.predict(test_x);
+		List<Double> predict_y=my_svm.predict(test_x);
 		
 		// 统计结果
 		int predict_error_counter = 0;
 		for(int i=0;i<test_number;i++){
-			float test_y_=test_y.get(i);
-			float predict_val=predict_y.get(i);
+			double test_y_=test_y.get(i);
+			double predict_val=predict_y.get(i);
 			if(Math.abs(predict_val-test_y_)>0.01){
 				System.out.format("[WRONG] predict=%f, test_y=%f\n",predict_val,test_y_);
 				predict_error_counter+=1;
@@ -44,7 +43,7 @@ public class Demo {
 		
 		// 正确率
 		System.out.format("train_num=%d, test_num=%d, predict_correct_num=%d\n",train_number, test_number, test_number - predict_error_counter);
-		System.out.format("predict accuracy=%.1f%%\n", (100 * (float)(test_number - predict_error_counter) / test_number));
+		System.out.format("predict accuracy=%.1f%%\n", (100 * (double)(test_number - predict_error_counter) / test_number));
 	}
 
 }
